feat: oplog

+++ calf-code-generator/src/main/java/com/github/shawven/calf/codeGenerator/mapper/base/BaseMapper.java
@@ -17,7 +17,7 @@ public interface BaseMapper<T> extends com.baomidou.mybatisplus.core.mapper.Base
         return selectOne(new QueryWrapper<>(entity));
     }
 
-    default Integer selectCount(T entity) {
+    default long selectCount(T entity) {
         return selectCount(new QueryWrapper<>(entity));
     }
